// src/runnersQcApp/shared/types.ts
export enum RecordStatusType {
  IDLE = "idle",
  IS_SYNCED = "isSynced",
  QUEUE = "queue",
}
// ...
export const RecTypeSchema = z.enum(
  Object.values(HorseMedicalRecType).filter(
    (value) => typeof value === "string"
  ) as [string, ...string[]]
);

export const RouteAdminSchema = z.enum(
  Object.values(HorseMedicalRouteAdmin).filter(
    (value) => typeof value === "string"
  ) as [string, ...string[]]
);

export const HorseRetireReasonSchema = z.enum(
  Object.values(HorseRetireReason).filter(
    (value) => typeof value === "string"
  ) as [string, ...string[]]
);

export const HorseDeathReasonSchema = z.enum(
  Object.values(HorseDeathReason).filter(
    (value) => typeof value === "string"
  ) as [string, ...string[]]
);

export const QCStatusSchema = z.enum(
  Object.values(QCStatus).filter((value) => typeof value === "string") as [
    string,
    ...string[]
  ]
);

export const HorseMedicalRecognizedNameSchema = z.object({
  name: z.string().nullable().optional(),
  expiry: z.string().optional(),
  administeredDate: z.string().optional(),
});

export const DocumentSchema = z.object({
  path: z.string(),
  id: z.string(),
  name: z.string(),
  type: z.string(),
  originalFile: z.any().optional(),
});

export const HorseMedicalResponseSchema = z.object({
  locationId: z.string().nullable().optional(),
  horseLocationId: z.string().nullable().optional(),
  hisaHorseId: z.string().min(1),
  designatedOwnerId: z.string().nullable().optional(),
  designatedOwner: z.string().nullable().optional(),
  responsibleHisaPersonId: z.string().nullable().optional(),
  createdBy: z.string().nullable().optional(),
  updatedBy: z.string().nullable().optional(),
  treatingHisaPersonId: z.string().nullable().optional(),
  attendingVetId: z.string().nullable().optional(),
  attendingVet: z.string().nullable().optional(),
  date: z.string().optional(),
  time: z.string().nullable().optional(),
  clearedToWork: z.boolean().optional(),
  clearedToRace: z.boolean().optional(),
  recType: RecTypeSchema,
  vaccine: z.string().nullable().optional(),
  drugName: z.string().nullable().optional(),
  drugRoute: RouteAdminSchema.optional(),
  drugDosage: z.string().nullable().optional(),
  procedure: z.string().nullable().optional(),
  surgery: z.string().nullable().optional(),
  dental: z.string().nullable().optional(),
  physiotherapy: z.string().nullable().optional(),
  chiropractic: z.string().nullable().optional(),
  modality: z.string().nullable().optional(),
  testName: z.string().nullable().optional(),
  necropsyCategory: z.string().nullable().optional(),
  testResults: z.string().nullable().optional(),
  limbTreated: z.string().nullable().optional(),
  notes: z.string().nullable().optional(),
  associatedDocumentUrls: z.array(z.string()).nullable().optional(),
  structure: z.string().nullable().optional(),
  description: z.string().nullable().optional(),
  relatedVetsListId: z.string().nullable().optional(),
  classifiedAs: z.string().nullable().optional(),
  retireReason: HorseRetireReasonSchema.nullable().optional(),
  dateOfDeath: z.string().nullable().optional(),
  deathReason: HorseDeathReasonSchema.nullable().optional(),
  latLng: z.string().nullable().optional(),
  drugPerscribingVetHisaId: z.string().nullable().optional(),
  treatingPersonRole: z.string().nullable().optional(),
  category: z.string().nullable().optional(),
  internalNotes: z.string().nullable().optional(),
  includeInStatus: z.boolean().nullable().optional(),
  qcStatus: QCStatusSchema.nullable().optional(),
  hisaHorseMedicalId: z.string().nullable().optional(),
  uuid: z.string().nullable().optional(),
  isDeleted: z.boolean().optional(),
  locationName: z.string().nullable().optional(),
  horseLocationName: z.string().nullable().optional(),
  hisaHorseName: z.string().nullable().optional(),
  responsibleHisaPersonName: z.string().nullable().optional(),
  treatingHisaPersonName: z.string().nullable().optional(),
  designatedOwnerName: z.string().nullable().optional(),
  attendingVetName: z.string().nullable().optional(),
  displayName: z.string().nullable().optional(),
  conditionTreated: z.string().nullable().optional(),
  recognizedVaccines: z
    .array(HorseMedicalRecognizedNameSchema)
    .nullable()
    .optional(),
  recognizedTests: z
    .array(HorseMedicalRecognizedNameSchema)
    .nullable()
    .optional(),
  //
  lastUpdate: z.string().optional(),
  createdDate: z.string().optional(),
  internalId: z.string().nullable().optional(),
  isSynced: z.boolean().optional(),
  timestamp: z.number().optional(),
  files: z.array(z.any()).optional(),
  documentPaths: z.array(DocumentSchema).optional(),
});

export type THorseMedicalRecord = z.infer<typeof HorseMedicalResponseSchema>;